+++ services/src/main/java/org/keycloak/services/resources/admin/SAMLFederationResource.java
@@ -287,7 +287,7 @@ public class SAMLFederationResource {
                 try {
                     realm.addIdentityProviderMapper(new IdentityProviderMapperModel(mapper, alias));
                 } catch (Exception e) {
-                    logger.info("Mapper with id = " + mapperId + " failed to add to Idp with alias = " + alias);
+                    logger.debug("Mapper with id = " + mapperId + " failed to add to Idp with alias = " + alias);
                 }
             });
         } else if (action.equals("update")) {
@@ -300,7 +300,7 @@ public class SAMLFederationResource {
                         realm.updateIdentityProviderMapper(idpMapper);
                     }
                 } catch (Exception e) {
-                    logger.info("Mapper with id = " + mapperId + " failed to update in the Idp with alias = " + alias);
+                    logger.debug("Mapper with id = " + mapperId + " failed to update in the Idp with alias = " + alias);
                 }
             });
         } else if (action.equals("remove")) {
@@ -310,7 +310,7 @@ public class SAMLFederationResource {
                     if (idpMapper != null)
                         realm.removeIdentityProviderMapper(idpMapper);
                 } catch (Exception e) {
-                    logger.info("Mapper with id = " + mapperId + " failed to remove from Idp with alias = " + alias);
+                    logger.debug("Mapper with id = " + mapperId + " failed to remove from Idp with alias = " + alias);
                 }
             });
         } else {